# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    Baseline:
      type: object
      properties:
        runId:
          type: string
        version:
          type: string
    Blocker:
      type: object
      properties:
        runId:
          type: string
        repoName:
          type: string
...
    CiBaselineAnswer:
      type: object
      properties:
        baseline:
          $ref: "#/components/schemas/CiReportBaselineDto"
    CiLiveStepDto:
      type: object
      properties:
        stepIndex:
          type: integer
          format: int32
        startedAt:
          $ref: "#/components/schemas/Instant"
        output:
          type: string
    CiReportBaselineDto:
      type: object
      properties:
        version:
          type: string
        runId:
          type: string
        releaseRequestId:
          type: string
        tagSha:
          type: string
    CiReportDto:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        kindVersion:
          type: integer
          format: int32
        stepIndex:
          type: integer
          format: int32
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/CiReportHighlightDto"
        baselineRunId:
          type: string
        baselineVersion:
          type: string
        payloadBytes:
          type: integer
          format: int32
        submittedAt:
          $ref: "#/components/schemas/Instant"
        payload:
          $ref: "#/components/schemas/JsonNode"
    CiReportHighlightDto:
      type: object
      properties:
        severity:
          type: string
        text:
          type: string
        metric:
          type: string
        value:
          type: number
          format: double
        delta:
          type: number
          format: double
    CiReportSubmission:
      type: object
      properties:
        kindVersion:
          type: integer
          format: int32
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/CiReportHighlightDto"
        baseline:
          $ref: "#/components/schemas/Baseline"
        payload:
          $ref: "#/components/schemas/JsonNode"
    CiReportSummaryDto:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        kindVersion:
          type: integer
          format: int32
        stepIndex:
          type: integer
          format: int32
        highlights:
          type: array
          items:
            $ref: "#/components/schemas/CiReportHighlightDto"
        baselineRunId:
          type: string
        baselineVersion:
          type: string
        payloadBytes:
          type: integer
          format: int32
        submittedAt:
          $ref: "#/components/schemas/Instant"
    CiRunDto:
      type: object
      properties:
        id:
          type: string
        repoId:
          type: string
        projectId:
          type: string
        repoName:
          type: string
        branch:
          type: string
        commitSha:
          type: string
        status:
          $ref: "#/components/schemas/CiRunStatus"
        createdAt:
          $ref: "#/components/schemas/Instant"
        startedAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
        cancellationReason:
          type: string
        supersededByRunId:
          type: string
        daemonVersion:
          type: string
        triggerType:
          $ref: "#/components/schemas/CiTriggerType"
        triggerEventId:
          type: string
        triggerEventName:
          type: string
        releaseRequestId:
          type: string
        retryOfRunId:
          type: string
        configPath:
          type: string
        archetypeName:
          type: string
        archetypeConfigPath:
          type: string
        archetypeRev:
          type: string
        archetypeVersion:
          type: string
        priority:
          type: string
        expectedStepDurationsMillis:
          type: array
          items:
            type: integer
            format: int64
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CiStepDto"
        live:
          $ref: "#/components/schemas/CiLiveStepDto"
        phase:
          $ref: "#/components/schemas/CiRunPhase"
        queuePosition:
          type: integer
          format: int32
        expectedStartInMillis:
          type: integer
          format: int64
        expectedFinishInMillis:
          type: integer
          format: int64
        predictionUnavailable:
          type: string
        ordering:
          $ref: "#/components/schemas/CiRunOrderingDto"
        runnerId:
          $ref: "#/components/schemas/UUID"
        runnerName:
          type: string
        retryReason:
          type: string
        autoRetry:
          type: boolean
    CiRunOrderingDto:
      type: object
      properties:
        position:
          type: integer
          format: int32
        kindTier:
          type: integer
          format: int32
        priority:
          type: string
        priorityRank:
          type: integer
          format: int32
        topologyBlockers:
          type: array
          items:
            $ref: "#/components/schemas/Blocker"
        selection:
          $ref: "#/components/schemas/Selection"
    CiRunPhase:
      type: string
      enum:
      - RELEASE_REQUEST
      - RELEASE
    CiRunReportsDto:
      type: object
      properties:
        runId:
          type: string
        commitSha:
          type: string
        releaseRequestId:
          type: string
        baseline:
          $ref: "#/components/schemas/CiReportBaselineDto"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/CiReportSummaryDto"
    CiRunStatus:
      type: string
      enum:
      - QUEUED
      - RUNNING
      - SUCCESS
      - FAILED
      - CANCELLED
      - CONFIG_ERROR
      - TIMED_OUT
...
    CiStepDto:
      type: object
      properties:
        stepIndex:
          type: integer
          format: int32
        image:
          type: string
        status:
          $ref: "#/components/schemas/CiStepStatus"
        exitCode:
          type: integer
          format: int32
        startedAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
        output:
          type: string
    CiStepStatus:
      type: string
      enum:
      - PENDING
      - RUNNING
      - SUCCESS
      - FAILED
      - SKIPPED
      - TIMED_OUT
    CiTriggerType:
      type: string
      enum:
      - POST_RECEIVE
      - EVENT
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    JsonNode:
      type: object
      properties:
        empty:
          type: boolean
        valueNode:
          type: boolean
        containerNode:
          type: boolean
        missingNode:
          type: boolean
        array:
          type: boolean
        object:
          type: boolean
        nodeType:
          $ref: "#/components/schemas/JsonNodeType"
        pojo:
          type: boolean
        number:
          type: boolean
        integralNumber:
          type: boolean
        floatingPointNumber:
          type: boolean
        short:
          type: boolean
        int:
          type: boolean
        long:
          type: boolean
        float:
          type: boolean
        double:
          type: boolean
        bigDecimal:
          type: boolean
        bigInteger:
          type: boolean
        textual:
          type: boolean
        boolean:
          type: boolean
        "null":
          type: boolean
        binary:
          type: boolean
    JsonNodeType:
      type: string
      enum:
      - ARRAY
      - BINARY
      - BOOLEAN
      - MISSING
      - "NULL"
      - NUMBER
      - OBJECT
      - POJO
      - STRING
...
    ListRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/CiRunDto"
...
    Selection:
      type: string
      enum:
      - UNBLOCKED
      - CYCLE_DEGRADED
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /ci/api/runs:
    get:
      summary: "List a repository's CI runs, newest first"
      operationId: listRuns
      parameters:
      - description: Return only the newest n runs; omit for all of them
        schema:
          type: integer
          minimum: 1
        name: limit
        in: query
      - description: The repository whose runs to list — required
        required: true
        name: repositoryId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: "The repository's runs with their step boundaries, without\
            \ step output"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRunsResponse"
        "400":
          description: "The repository id is missing or invalid, or the limit is not\
            \ a positive integer"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Ci Run Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
...
  /ci/api/runs/{runId}/baseline:
    get:
      summary: "A run's baseline: the gating run of its repository's newest released\
        \ version, or null"
      operationId: getRunBaseline
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "The baseline, or {\"baseline\":null}"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CiBaselineAnswer"
        "404":
          description: No such run
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Ci Report Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
        - qits:ci-run
  /ci/api/runs/{runId}/baseline/reports/{kind}:
    get:
      summary: "The baseline run's release reports of one kind, with payloads"
      operationId: listRunBaselineReports
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          type: string
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "The reports; [] when there is no baseline or none"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CiReportDto"
        "400":
          description: A bad kind name
        "404":
          description: No such run
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Ci Report Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
        - qits:ci-run
...
  /ci/api/runs/{runId}/reports:
    get:
      summary: "A run's release reports, without payloads, and its baseline"
      operationId: listRunReports
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The run's reports; an empty list when it has none
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CiRunReportsDto"
        "404":
          description: No such run
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Ci Report Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
        - qits:ci-run
  /ci/api/runs/{runId}/reports/{reportId}:
    get:
      summary: "One release report of a run, with its payload"
      operationId: getRunReport
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CiReportDto"
        "404":
          description: "No such run, or no such report on it"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Ci Report Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
        - qits:ci-run
...
  /ci/api/runs/{runId}/steps/{stepIndex}/reports/{kind}:
    put:
      summary: Submit a step's release report of one kind (a run's own ci-run token
        only)
      operationId: putRunStepReport
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          type: string
      - name: runId
        in: path
        required: true
        schema:
          type: string
      - name: stepIndex
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/CiReportSubmission"
        required: true
      responses:
        "204":
          description: "Stored, replacing any earlier report of the kind"
        "400":
          description: "A bad kind name, body, or more than 10 highlights"
        "403":
          description: The token is not this run's
        "404":
          description: No such run
        "409":
          description: The run is not RUNNING
        "413":
          description: The body is larger than 1 MiB
        "401":
          description: Not Authorized
      tags:
      - Ci Report Controller
      security:
      - SecurityScheme:
        - qits:ci-run
info:
  title: qits-ci API
  version: 2026.1010.81424